Add outline of Docker integration test runner
