+++ kafka2hbase/__init__.py
@@ -5,12 +5,13 @@ from . import shovel_stream as _shovel
 from . import kafka as _kafka
 from . import hbase as _hbase
 from . import message as _message
+from . import transform as _transform
 
 load_config = _env.load_config
 all_env_vars = _env.all_env_vars
 
 shovel = _shovel.shovel
-qualified_table_name = _shovel.qualified_table_name
+qualified_table_name = _transform.qualified_table_name
 kafka_stream = _kafka.kafka_stream
 hbase_put = _hbase.hbase_put
