+++ requirements-dev.txt
@@ -2,5 +2,4 @@ docker-compose==1.24.0
 invoke==1.2.0
 pytz==2019.1
 nose==1.3.7
-behave==1.2.6
 assertpy==0.14
